chore(miniapp): add react config
