+++ miniapp/package.json
@@ -6,7 +6,14 @@
   "scripts": {
     "build": "vite build"
   },
+  "dependencies": {
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
+  },
   "devDependencies": {
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@vitejs/plugin-react-swc": "^3.5.0",
     "vite": "^5.4.1"
   }
 }
