+++ miniapp/src/App.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default function App() {
+  return (
+    <div className="card coming-soon">
+      <h1>Dynamic Capital VIP</h1>
+      <p className="muted">Our mini app is coming soon. Stay tuned!</p>
+    </div>
+  );
+}
