+++ miniapp/vite.config.ts
@@ -1,6 +1,8 @@
 import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react-swc';
 
 export default defineConfig({
+  plugins: [react()],
   base: '/assets/',
   build: {
     outDir: 'dist',
